+++ src/ultimate_notion/__init__.py
@@ -48,7 +48,7 @@ from ultimate_notion.blocks import (
 from ultimate_notion.core import Workspace, WorkspaceType, get_active_session
 from ultimate_notion.database import Database
 from ultimate_notion.emoji import Emoji
-from ultimate_notion.file import FileInfo
+from ultimate_notion.file import AnyFile, ExternalFile, NotionFile, url
 from ultimate_notion.obj_api.enums import AggFunc, BGColor, CodeLang, Color, NumberFormat, OptionGroupType, VState
 from ultimate_notion.option import Option, OptionGroup, OptionNS
 from ultimate_notion.page import Page
@@ -62,6 +62,7 @@ from ultimate_notion.utils import DateTimeOrRange, SList
 __all__ = [
     'PDF',
     'AggFunc',
+    'AnyFile',  # for type hinting only
     'BGColor',
     'Block',  # for type hinting only
     'Bookmark',
@@ -80,14 +81,15 @@ __all__ = [
     'Embed',
     'Emoji',
     'Equation',
+    'ExternalFile',
     'File',
-    'FileInfo',
     'Heading1',
     'Heading2',
     'Heading3',
     'Image',
     'LinkPreview',
     'LinkToPage',
+    'NotionFile',
     'NumberFormat',
     'NumberedItem',
     'Option',
@@ -121,4 +123,5 @@ __all__ = [
     'mention',
     'prop',
     'text',
+    'url',
 ]