+++ tests/test_file.py
@@ -5,9 +5,14 @@ import pytest
 import ultimate_notion as uno
 
 
-@pytest.mark.vcr()
-def test_file_upload(notion: uno.Session) -> None:
+@pytest.mark.file_upload
+def test_file_upload(root_page: uno.Page, notion: uno.Session) -> None:
     with open('docs/assets/images/favicon.png', 'rb') as file:
         file_info = notion.upload(file=file)
 
-    assert file_info is not None
+    page = notion.create_page(parent=root_page, title='Test Page to check file upload as icon')
+    page.icon = file_info
+
+    # from IPython import embed
+
+    # embed()  # for debugging